requirements.txt
aiogram==3.8.0
gspread==6.1.2
redis==5.0.7
python-dotenv==1.0.1
oauth2client==4.1.3
